backend/requirements.txt
# Core Flask and web framework dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database drivers
# SQLite (built-in with Python)
# For Azure SQL Database
pyodbc==5.0.1
# Alternative SQL Server driver
pymssql==2.2.8

# Authentication and security
cryptography==41.0.7
PyJWT==2.8.0

# Date and time handling
pytz==2023.3

# File handling and media processing
Pillow==10.0.1
reportlab==4.0.4

# HTTP and API utilities
requests==2.31.0

# AI and machine learning (for audio processing)
openai==1.3.5

# Audio processing (optional, for advanced audio features)
pydub==0.25.1

# Environment and configuration
python-dotenv==1.0.0

# Production server (for Azure)
gunicorn==21.2.0
waitress==2.1.2

# Development and debugging tools (optional for production)
python-decouple==3.8

# Azure-specific packages
azure-storage-blob==12.18.3
azure-identity==1.15.0

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0

# Logging and monitoring
structlog==23.2.0